Service for Maps Personalization
This is part of what will later be a fairly large service to detect
personalization in online maps services. This service can fetch and
save images from Google Maps. Future plans are to add more map
providers, do image processing, set up cron jobs to fetch and process
images, set up a RESTful api for interacting with the data, set up a database
to store the images, and develop a front end.