add Map model
This is the model that represents each map that is fetched. It holds a MapRequest
which is all the unique data for the request to the maps api, as well as a date
on when the request was made, a boolean if the image has changed, and possibly
an image.