add a FetchJob table
The FetchJob table is used to keep track how many times the job has been run and when each one
occurred. This will allow us to map each Map row to a job execution. With this we can compare one
jobs images to another jobs images to check for personalization.