add passedControl to submit form
Apparently you have to submit an answer to mturk. This is the type of
shit that should be documented in BOLD. It should be clear that if this
is missing, then something wont work. And even if mturk felt the need that
it doesnt have to be documented, they should give a goddamn meaningful
error response rather than just "there was an error submitting your HIT".
Rant over.

Either way, this is a good thing to include because it will make it easier
for us to see if the turker passed the control.