add healthcheck to maintain db connections
This is hacky, but it is the best solution I have at the moment.
The problem is that if noone makes a request for a certain amount
of time, the db connection pool will get killed. This ensures that
a request is made every 10 minutes to maintain the db connections