some serious cleanup
Over the last 9 months I have yet to clean up unused code, which there is a lot
of. This is the first of what might be many cleanups. I basically just started
at a file that I knew wasnt used and worked backwards to every old import reference.
It is worth noting some of this code was used, like all of the monitoring code.
I dont really use that anymore because the crawler is quite stable and I have more
time to devote to the project