first part of major refactoring
These changes address the performance and stability related to mysql
connections and overall design. This should make it easier to fix
changes in the future. This is not quite done though. The only part
fully functional is the web server. The crawler still needs to be
interfaced with the CLI