part 1 of bing amt/admin integration
This is the first commit(of what might be many) to get Bing fully integrated.
This adds the exisitng functionality that google has to the service and admin UI.
This needs some serious testing before any HITs are sent out.
I also need to bring back the update job which creates hits based off historical
crawls in order to create the HITs